Non shared changes for StringBuilder
